Move SecurityUtils methods closer to the usage
